requirements.txt
accelerate==0.28.0


huggingface-hub==0.20.3


joblib==1.3.2

langchain==0.0.184
langchain-community==0.0.20
langchain-core==0.1.23

langchain-pinecone==0.0.2
langchainplus-sdk==0.0.20
langdetect==1.0.9
langsmith==0.0.87
layoutparser==0.3.4
lxml==4.9.4

altair==4.0
matplotlib==3.7.5

nltk==3.8.1
numexpr==2.8.6
numpy==1.24.4


opencv-python==4.9.0.80

pandas==2.0.3

pdf2image==1.17.0
pdfplumber==0.10.4

pikepdf==8.13.0
pillow==10.2.0
pinecone-client==3.0.3




pydantic==1.10.14



pyparsing==3.1.1
pypdf==4.0.1
PyPDF2==3.0.1
pypdfium2==4.27.0
PyPika==0.48.9


python-dotenv==1.0.0


PyYAML==6.0.1

regex==2023.12.25
requests==2.31.0

scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.2.2

streamlit==1.18.1
streamlit-chat==0.1.1

tiktoken==0.4.0
timm==0.9.12
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1

torch==2.2.0

tqdm==4.66.2

transformers==4.37.2


uvicorn==0.29.0

websocket-client==1.7.0
websockets==12.0


